ParticleLevelSet3: better surface track
- Tracking the surface is now done taking phi value into consideration
- CFL computation now in parallel
- Velocities attribution now in parallel
- Added support to log all timestamps on Timer class

+++ leviathan/fluids/levelset_fluid3.h
@@ -125,7 +125,7 @@ public:
    */
   std::vector<int> trackSurface();
 
-  std::vector<int> findSurfaceCells(int distacenToSurface);
+  std::vector<int> findSurfaceCells(double distacenToSurface);
   std::vector<int> findSurfaceCells();
 
 protected:
